pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-bigquery-server"
version = "0.1.0"
description = "A FastMCP server for securely accessing BigQuery datasets with support for HTTP and Stdio transport."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastmcp>=0.1.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "google-cloud-bigquery>=3.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase", # Added Supabase
    "postgrest>=1.0.2",
    "streamlit>=1.36.0",
    "openai>=1.30.0",
    "anthropic>=0.32.0",
    "google-generativeai>=0.7.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
mcp-bigquery = "mcp_bigquery.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_bigquery"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/mcp_bigquery --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
